requirements.txt
joblib==1.4.2
scikit-learn==1.3.0
rpy2==3.5.17
tqdm==4.67.1
pandas==2.2.3
biopython==1.85
streamlit==1.42.1
numpy==1.26.4
lightgbm==4.5.0
